requirements.txt
appdirs==1.4.3
attrs==18.2.0
behave==1.2.6
cached-property==1.5.1
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
idna==2.7
isodate==0.6.0
lxml==4.2.5
parse==1.9.0
parse-type==0.4.2
pytz==2018.5
PyYAML==3.13
requests==2.20.0
requests-toolbelt==0.8.0
ruamel.yaml==0.15.74
six==1.11.0
urllib3==1.24
zeep==3.1.0
